python/03-djangocms/requirements/requirements.txt
Django==1.4.1
Jinja2==2.6
Pygments==1.5
Sphinx==1.1.3
docutils==0.9.1
requests==0.13.6
wsgiref==0.1.2
django-staticfiles==1.2.1
South==0.7.6
django-cms==2.3.1
django-filer
pystache